refactor(e2e): Extract fixture e2e sections into components
