+++ integration/templates/expo-native/App.tsx
@@ -1,10 +1,12 @@
-import { ClerkProvider, useAuth, useSignIn, useUser } from '@clerk/expo';
+import { ClerkProvider, useAuth, useUser } from '@clerk/expo';
 import { useSignInWithGoogle } from '@clerk/expo/google';
 import { AuthView, UserButton } from '@clerk/expo/native';
 import { tokenCache } from '@clerk/expo/token-cache';
-import { requireOptionalNativeModule } from 'expo';
 import { useState } from 'react';
-import { Button, Modal, StyleSheet, Text, TextInput, View } from 'react-native';
+import { Button, Modal, StyleSheet, Text, View } from 'react-native';
+
+import { E2EControls } from './components/E2EControls';
+import { JsSignInForm } from './components/JsSignInForm';
 
 const publishableKey = process.env.EXPO_PUBLIC_CLERK_PUBLISHABLE_KEY;
 
@@ -12,57 +14,13 @@ if (!publishableKey) {
   throw new Error('Missing EXPO_PUBLIC_CLERK_PUBLISHABLE_KEY');
 }
 
-// Fixture-local Maestro hook (modules/e2e-hooks); android-only, null elsewhere.
-const E2EHooks = requireOptionalNativeModule<{ corruptNativeDeviceToken(): Promise<boolean> }>('E2EHooks');
-
 function NativeBuildFixture() {
-  const { isLoaded, isSignedIn, getToken, signOut } = useAuth({ treatPendingAsSignedOut: false });
+  const { isLoaded, isSignedIn, signOut } = useAuth({ treatPendingAsSignedOut: false });
   const { user } = useUser();
-  const { signIn } = useSignIn();
   const { startGoogleAuthenticationFlow } = useSignInWithGoogle();
   const [isAuthOpen, setIsAuthOpen] = useState(false);
   const [googleResult, setGoogleResult] = useState<string | null>(null);
-  const [identifier, setIdentifier] = useState('');
-  const [password, setPassword] = useState('');
   const [e2eStatus, setE2eStatus] = useState<string | null>(null);
-
-  const onJsSignIn = async () => {
-    try {
-      const { error } = await signIn.password({ identifier, password });
-      if (error) {
-        setE2eStatus(`sign-in-error ${error.message ?? ''}`.replace(/\s+/g, ' ').trim());
-        return;
-      }
-      const { error: finalizeError } = await signIn.finalize();
-      if (finalizeError) {
-        setE2eStatus(`sign-in-error ${finalizeError.message ?? ''}`.replace(/\s+/g, ' ').trim());
-      }
-    } catch (error) {
-      setE2eStatus(`sign-in-error ${String(error)}`.replace(/\s+/g, ' '));
-    }
-  };
-
-  const onCorruptNativeToken = async () => {
-    setE2eStatus(null);
-    try {
-      const didCorrupt = await E2EHooks?.corruptNativeDeviceToken();
-      // Delay the marker so Maestro cannot race the native client event and
-      // the JS sync settling.
-      setTimeout(() => setE2eStatus(didCorrupt ? 'corrupt-done' : 'corrupt-failed'), 3000);
-    } catch {
-      setE2eStatus('corrupt-failed');
-    }
-  };
-
-  const onMintSessionToken = async () => {
-    setE2eStatus(null);
-    try {
-      const token = await getToken({ skipCache: true });
-      setE2eStatus(token ? 'token-ok' : 'token-empty');
-    } catch {
-      setE2eStatus('token-error');
-    }
-  };
 
   return (
     <View style={styles.container}>
@@ -91,48 +49,8 @@ function NativeBuildFixture() {
         />
       )}
       {googleResult && <Text testID='google-result'>{googleResult}</Text>}
-      {!isSignedIn && (
-        <>
-          <TextInput
-            testID='e2e-identifier-input'
-            style={styles.input}
-            autoCapitalize='none'
-            autoCorrect={false}
-            placeholder='e2e identifier'
-            value={identifier}
-            onChangeText={setIdentifier}
-          />
-          <TextInput
-            testID='e2e-password-input'
-            style={styles.input}
-            autoCapitalize='none'
-            autoCorrect={false}
-            secureTextEntry
-            placeholder='e2e secret'
-            value={password}
-            onChangeText={setPassword}
-          />
-          <Button
-            testID='e2e-js-sign-in-button'
-            title='E2E JS sign in'
-            onPress={() => void onJsSignIn()}
-          />
-        </>
-      )}
-      {isSignedIn && (
-        <View style={styles.e2eRow}>
-          <Button
-            testID='e2e-corrupt-native-token-button'
-            title='Corrupt'
-            onPress={() => void onCorruptNativeToken()}
-          />
-          <Button
-            testID='e2e-refresh-token-button'
-            title='Mint'
-            onPress={() => void onMintSessionToken()}
-          />
-        </View>
-      )}
+      {!isSignedIn && <JsSignInForm onStatus={setE2eStatus} />}
+      {isSignedIn && <E2EControls onStatus={setE2eStatus} />}
       {e2eStatus && <Text testID='e2e-status'>{e2eStatus}</Text>}
       {isSignedIn && (
         <Button
@@ -172,23 +90,11 @@ const styles = StyleSheet.create({
     justifyContent: 'center',
     padding: 24,
   },
-  e2eRow: {
-    flexDirection: 'row',
-    gap: 12,
-    justifyContent: 'space-between',
-  },
   header: {
     alignItems: 'center',
     flexDirection: 'row',
     gap: 12,
     justifyContent: 'space-between',
-  },
-  input: {
-    borderColor: '#cccccc',
-    borderRadius: 6,
-    borderWidth: 1,
-    paddingHorizontal: 12,
-    paddingVertical: 8,
   },
   title: {
     flex: 1,
